// src/app/bookings/domain/enums/nextDaysEnum.ts
export enum NextDaysEnum {
  TODAY = 0,
  NEXT_7_DAYS = 7,
  NEXT_15_DAYS = 15,
  NEXT_30_DAYS = 30,
  NEXT_90_DAYS = 90,
  NEXT_180_MONTHS = 180
}

export enum NextDaysIDEnum {
  EMPTY = 'empty',
  TODAY = 'today',
  NEXT_7_DAYS = 'next7Days',
  NEXT_15_DAYS = 'next15Days',
  NEXT_30_DAYS = 'next30Days',
  NEXT_90_DAYS = 'next90Days',
  NEXT_180_MONTHS = 'next180Months'
}