// src/app/bookings/domain/enums/drawerStep.ts
export enum DrawerSteps {
  DETAIL = 'detail',
  SEND_CONFIRMATION = 'sendConfirmation',
  CANCEL_RESERVATION = 'cancelReservation'
}

export enum DrawerStepsInSendConfirmation {
  FORM = 'formSendConfirmartion',
  SUCCESS_MESSAGE = 'successMessage'
}

export enum DrawerStepsInCancelReservation {
  CANCEL_RESERVATION = 'cancelReservation',
  SUCCESS_MESSAGE = 'successMessage'
}